Update SM-portspeed.sh
Found differences between the Estoc and Sabre platforms in regards to Node Number sequence so made 2 separate array values. Still haven't put in the logic yet that detects which is which (primarily because I don't have an Opteron chassis to test with) but will add it in when the opportunity arises.